Move UI card service in card repository
